Changed package name using naming conventions
