+++ src/main/java/za/ac/cput/entity/CarLot.java
@@ -1,4 +1,4 @@
-package za.ac.cput.Entity;
+package za.ac.cput.entity;
 
 /*
    @Description:CarLot.java
