+++ src/main/java/za/ac/cput/entity/Client.java
@@ -1,4 +1,4 @@
-package za.ac.cput.Entity;
+package za.ac.cput.entity;
 
 /*
    @Description:Client.java -> Client class for Client Entity
