+++ src/main/java/za/ac/cput/entity/ClientAccount.java
@@ -1,4 +1,4 @@
-package za.ac.cput.Entity;
+package za.ac.cput.entity;
 /*
     @Description: ClientAccount Entity ->
     @Author: Asiphiwe Hanjiwe
