+++ src/main/java/za/ac/cput/factory/ClientFactory.java
@@ -1,4 +1,4 @@
-package za.ac.cput.Factory;
+package za.ac.cput.factory;
 
 /*
    @Description:ClientFactory.java
@@ -7,8 +7,8 @@ package za.ac.cput.Factory;
    @Date: 3 June 2021
  */
 
-import za.ac.cput.Entity.Client;
-import za.ac.cput.Util.GenericHelper;
+import za.ac.cput.entity.Client;
+import za.ac.cput.util.GenericHelper;
 
 public class ClientFactory {
     public static Client createClient(String name, String surname, String contactNumber,String accountNumber) {
