+++ src/main/java/za/ac/cput/factory/EmployeeFactory.java
@@ -1,12 +1,12 @@
-package za.ac.cput.Factory;
+package za.ac.cput.factory;
 /*
     @Description:Employees Entity ->
     @Author: Tyronne Lloyd Hendricks
     @Student Number: 215141210
     @Date: 3rd June 2021
  */
-import za.ac.cput.Entity.Employee;
-import za.ac.cput.Util.GenericHelper;
+import za.ac.cput.entity.Employee;
+import za.ac.cput.util.GenericHelper;
 
 public class EmployeeFactory {
     public static Employee createEmployee(String name, String surname,String userType, String username,String password){
