+++ src/main/java/za/ac/cput/factory/LoginFactory.java
@@ -1,11 +1,11 @@
-package za.ac.cput.Factory;
+package za.ac.cput.factory;
 /*
     @Description: Login Factory ->
     @Author: Asiphiwe Hanjiwe
     @Student Number: 218336675
     @Date: 04 June 2021
  */
-import za.ac.cput.Entity.Login;
+import za.ac.cput.entity.Login;
 
 public class LoginFactory {
     public static Login createLogin(String userType,String password, String username)
