+++ src/test/java/za/ac/cput/factory/CarLotFactoryTest.java
@@ -1,4 +1,4 @@
-package za.ac.cput.Factory;
+package za.ac.cput.factory;
 
 /*
    @Description:CarLotFactory.java
@@ -6,7 +6,7 @@ package za.ac.cput.Factory;
    @Student Number: 218150040
    @Date: 4 June 2021
  */
-import za.ac.cput.Entity.CarLot;
+import za.ac.cput.entity.CarLot;
 import org.junit.jupiter.api.*;
 import static org.junit.jupiter.api.Assertions.*;
