+++ src/test/java/za/ac/cput/factory/CarRentalFactoryTest.java
@@ -1,4 +1,4 @@
-package za.ac.cput.Factory;
+package za.ac.cput.factory;
 
 /*
     @Description: CarRentalFactoryTest.java
@@ -10,7 +10,7 @@ package za.ac.cput.Factory;
 import org.junit.jupiter.api.Disabled;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.Timeout;
-import za.ac.cput.Entity.CarRental;
+import za.ac.cput.entity.CarRental;
 
 import static org.junit.jupiter.api.Assertions.*;
